refactor(session-invalidation) Refactors session invalidation to User.flush() API
