+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "authentication-client",
-  "version": "1.7.1",
+  "version": "1.8.0",
   "description": "A thin Authentication API consumer",
   "homepage": "https://github.com/Avocarrot/authentication-client",
   "main": "src/index.js",
