+++ src/models/session.js
@@ -1,6 +1,5 @@
 const assert = require('assert');
 const User = require('../models/user');
-const Store = require('../services/store');
 const retrieveURL = require('../utils').retrieveURL;
 const redirectToURL = require('../utils').redirectToURL;
 
@@ -13,7 +12,6 @@ class Session {
    * Initializes User
    *
    * @constructor
-   * @param {Store} store - The Store instance to use
    * @param {User} consumer - The User instance to use
    * @param {String} loginHost - The login app host
    * @param {String} redirectFn - The function the forces URL redirection - Defaults to `window.location.replace`
@@ -21,11 +19,9 @@ class Session {
    * @return {User}
    *
    */
-  constructor(store, user, loginHost, redirectFn = redirectToURL, pageURL = retrieveURL) {
-    assert(store instanceof Store, '`store` should be instance of Store');
+  constructor(user, loginHost, redirectFn = redirectToURL, pageURL = retrieveURL) {
     assert(user instanceof User, '`user` should be instance of User');
     assert(loginHost, '`loginHost` is not defined');
-    this._store = store;
     this._user = user;
     this._loginHost = loginHost;
     this._redirectFn = redirectFn;
@@ -49,9 +45,6 @@ class Session {
    *
    */
   invalidate() {
-    // Flush stored data
-    this._store.remove('access_token');
-    this._store.remove('refresh_token');
     // Redirect to login host with a return URL
     return this._redirectFn(`${this._loginHost}/login`);
   }
